Making service URL configurable for dev and test
The ZIPKIN_SERVICE_URL environment variable was hardcoded to localhost
for the dev and test environments.

This causes issues when for instance running your application in a container
as localhost will resolve to the container in which the application is running,
which in most cases won't actually be where the zipkin service is running.